Completed UI for login and register UI screen
