+++ client/app/auth/login.tsx
@@ -5,17 +5,18 @@ import { ThemedView } from '@/components/ThemedView';
 import { useForm } from 'react-hook-form';
 import InputBox from '@/components/InputBox';
 import { SafeAreaView } from 'react-native-safe-area-context';
+import { Link } from 'expo-router';
 
 
-const Signup = () => {
-  const { control, handleSubmit, formState: { errors , isValid } } = useForm({
+const Login = () => {
+  const { control, handleSubmit, formState: { errors, isValid } } = useForm({
     defaultValues: {
       email: '',
       password: '',
       confirmPassword: ''
     },
     mode: "onChange"
-        
+
   });
 
   const onSubmit = (data) => {
@@ -27,10 +28,7 @@ const Signup = () => {
       <ThemedView style={styles.container}>
 
         <ThemedView style={styles.header}>
-          <ThemedText type={'title'} style={styles.title}>Create Account</ThemedText>
-          <ThemedText style={styles.subtitle}>
-            Create an account to start chatting with friends and explore the community!
-          </ThemedText>
+          <ThemedText type={'title'} style={styles.title}>Login to your account</ThemedText>
         </ThemedView>
 
         <ThemedView style={styles.form}>
@@ -68,28 +66,22 @@ const Signup = () => {
             }}
             error={errors.password}
           />
-
-          <InputBox
-            control={control}
-            name="confirmPassword"
-            label="Confirm Password"
-            placeholder="Confirm your password"
-            rules={{
-              required: 'Please confirm your password',
-              validate: (value, formValues) =>
-                value === formValues.password || 'Passwords do not match'
-            }}
-            error={errors.confirmPassword}
-          />
         </ThemedView>
 
         <Pressable
-          style={!isValid ?[styles.button , styles.buttonDisabled] : styles.button}
+          style={!isValid ? [styles.button, styles.buttonDisabled] : styles.button}
           onPress={handleSubmit(onSubmit)}
           disabled={!isValid}
         >
-          <ThemedText style={styles.buttonText}>Sign up</ThemedText>
+          <ThemedText style={styles.buttonText}>Login</ThemedText>
         </Pressable>
+
+
+        <Link href="/auth/register" style={styles.noAccountTextContainer}><ThemedText style={styles.noAccountText}>
+          Do not have a account?
+        </ThemedText>
+        </Link>
+
       </ThemedView>
 
     </SafeAreaView>
@@ -130,10 +122,20 @@ const styles = StyleSheet.create({
     fontSize: 16,
     fontWeight: '600',
   },
-    buttonDisabled: {
-    backgroundColor: '#D1D5DB', 
+  buttonDisabled: {
+    backgroundColor: '#D1D5DB',
   },
-    
+  noAccountTextContainer: {
+    textAlign: "center",
+    paddingVertical: 10,
+  },
+
+  noAccountText: {
+    textAlign: "center",
+  }
 });
 
-export default Signup;
+export default Login;
+
+
+
